マルチスレッド対応と描画ロジックの最適化
`GameRunner` クラスにマルチスレッド対応を追加し、更新処理と描画処理を分離。`NowFPS` の計算ロジックを改善し、`Platform.SystemFPS` を活用。`TargetFps` の型変更やフォント描画処理のリファクタリングを実施。

その他、FPS 表示位置の調整、キー入力やムービー再生ロジックの整理、スレッドセーフな同期処理の追加など、全体的なコードの最適化と可読性向上を実施。

+++ Core/Config.cs
@@ -13,7 +13,7 @@ public sealed class GameConfig
     public bool ShowMouse { get; set; } = true;
 
     // --- Timing / Performance ---
-    public float TargetFps { get; set; } = 60f;
+    public int TargetFps { get; set; } = 60;
     public bool VSync { get; set; } = false;
     public bool UseMultiThreadUpdate { get; set; } = false;
     public int SleepDurationMs { get; set; } = 1000 * 60 * 10; // 長時間放置でスリープするまで
